requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mkdocs>=1.6.0
mkdocs-material>=9.5.0
pymdown-extensions>=10.0.0
